Out with the old, in with the new
This commit is long overdue. The main change is to upgrade shoulda's
hard dependency on shoulda-matchers from 2.x to 3.x. Since this by
itself is a pretty consequential change, though, I also took the
opportunity to make some more at the same time that I think are
reasonable:

* Remove support for Rails 3, Rails 4.0, and Rails 4.1, since they have
  been end-of-lifed; support Rails 4.2-5.0 instead.
* Remove support for Ruby 1.9 and 2.0; support Ruby 2.1-2.4 instead.
* Remove support for RSpec, along with related Cucumber specs. People
  who use RSpec won't be using the `shoulda` gem -- they'll be using
  `shoulda-matchers` instead. There's already a good collection of
  integration specs there, so we don't need the tests here.

Please note that we are using the Rails Git repo to refer to 4.2 because
the latest release of 4.2 isn't compatible with Ruby 2.4 just yet (but
it's fixed on the 4-2-stable branch).