Replace Cucumber tests with RSpec tests
Copy over the acceptance test framework that we are using for
shoulda-matchers. This is superior because as a part of running the
acceptance, we execute commands and we have it so that if one of those
commands fails, the output from that command is printed to the console
so that we can better debug the issue. This is something that Cucumber
(via Aruba) does not do.