Upgrade all the things, 2023 edition
This gem is severely out of date. Not only is the version of
`shoulda-matchers` very old, but I can't even get this project up and
running on my computer anymore because it relies on an old version of
Ruby.

So, the primary motivation behind this commit is to bump the dependency
on `shoulda-matchers` to 5.x.

Besides this, in `shoulda-matchers` there are changes that either have
been merged or are in flight which aim to upgrade dependencies, drop
support for old Rubies and Rails, etc., and I've integrated them into
this commit as well. These include:

* Update development Ruby to 3.2.1
* Replace Travis config with GitHub Actions workflows
* Upgrade Rubocop configuration, and correct lint violations
* Add support for Ruby 3.0, 3.1, and 3.2 as well as Rails 6.1 and 7.0
* Drop support for Ruby < 3 and Rails < 6.1

In addition, I've had to update some of the tests which failed after
performing these upgrades:

* Patch Snowglobe (which is used in tests to generate a Rails
  application) due to some incompatibilities with recent Rails versions
  and bugs
* Make sure to install `rails-controller-testing` and `bcrypt` in
  generated Rails application
* Update tests involving `filter_param` as Rails changed which
  parameters are filtered in recent versions